Seperated Follower and FollowerCalculator
