+++ src/main/java/com/pedropathing/localization/PoseUpdater.java
@@ -25,8 +25,6 @@ import com.pedropathing.pathgen.Vector;
  * @version 1.0, 3/4/2024
  */
 public class PoseUpdater {
-    private HardwareMap hardwareMap;
-
     private IMU imu;
 
     private Localizer localizer;
@@ -51,17 +49,15 @@ public class PoseUpdater {
     private long currentPoseTime;
 
     /**
-     * Creates a new PoseUpdater from a HardwareMap and a Localizer.
+     * Creates a new PoseUpdater from a Localizer and sets constants.
      *
-     * @param hardwareMap the HardwareMap
      * @param localizer the Localizer
      * @param FConstants the constants for the Follower
      * @param LConstants the constants for the Localizer
      */
-    public PoseUpdater(HardwareMap hardwareMap, Localizer localizer, Class<?> FConstants, Class<?> LConstants) {
+    public PoseUpdater(Localizer localizer, Class<?> FConstants, Class<?> LConstants) {
         Constants.setConstants(FConstants, LConstants);
 
-        this.hardwareMap = hardwareMap;
         this.localizer = localizer;
 
         if (localizer.getClass() != PinpointLocalizer.class) {
@@ -75,24 +71,22 @@ public class PoseUpdater {
     }
 
     /**
-     * Creates a new PoseUpdater from a HardwareMap.
+     * Creates a new PoseUpdater from a HardwareMap and sets constants
      *
      * @param hardwareMap the HardwareMap
      * @param FConstants the constants for the Follower
      * @param LConstants the constants for the Localizer
      */
     public PoseUpdater(HardwareMap hardwareMap, Class<?> FConstants, Class<?> LConstants) {
-        this(hardwareMap, createLocalizer(hardwareMap), FConstants, LConstants);
+        this(createLocalizer(hardwareMap), FConstants, LConstants);
     }
 
     /**
-     * Creates a new PoseUpdater from a HardwareMap and a Localizer.
+     * Creates a new PoseUpdater from a Localizer.
      *
-     * @param hardwareMap the HardwareMap
      * @param localizer the Localizer
      */
-    public PoseUpdater(HardwareMap hardwareMap, Localizer localizer) {
-        this.hardwareMap = hardwareMap;
+    public PoseUpdater(Localizer localizer) {
         this.localizer = localizer;
 
         if (localizer.getClass() != PinpointLocalizer.class) {
@@ -111,7 +105,19 @@ public class PoseUpdater {
      * @param hardwareMap the HardwareMap
      */
     public PoseUpdater(HardwareMap hardwareMap) {
-        this(hardwareMap, createLocalizer(hardwareMap));
+        this(createLocalizer(hardwareMap));
+    }
+
+    /**
+     * Creates a new PoseUpdater from a HardwareMap and a Localizer. Note:
+     * this constructor is there for backwards compatibility, because
+     * hardwareMap is not needed when giving a localizer. This can be removed
+     * in the next major release.
+     *
+     * @param hardwareMap the HardwareMap
+     */
+    public PoseUpdater(HardwareMap hardwareMap, Localizer localizer) {
+        this(localizer);
     }
 
     private static Localizer createLocalizer(HardwareMap hardwareMap) {
